+++ Tests/Unit/GameAction/Util/GameAction_Player.cs
@@ -33,6 +33,12 @@ namespace Helion.Tests.Unit.GameAction
         public static void PlayerRunBackward(SinglePlayerWorld world, double angle, Func<bool> runWhile, TimeSpan? timeout = null, Action? onTick = null) =>
             RunPlayerCommands(world, angle, [TickCommands.Backward], runWhile, timeout, onTick, stopTicks: null, tickAngleTurn: null);
 
+        public static void PlayerStrafeLeft(SinglePlayerWorld world, double angle, Func<bool> runWhile, TimeSpan? timeout = null, Action? onTick = null) =>
+            RunPlayerCommands(world, angle, [TickCommands.Left], runWhile, timeout, onTick, stopTicks: null, tickAngleTurn: null);
+
+        public static void PlayerStrafeRight(SinglePlayerWorld world, double angle, Func<bool> runWhile, TimeSpan? timeout = null, Action? onTick = null) =>
+            RunPlayerCommands(world, angle, [TickCommands.Right], runWhile, timeout, onTick, stopTicks: null, tickAngleTurn: null);
+
         public static void RunPlayerCommands(SinglePlayerWorld world, double angle, TickCommands[] commands, Func<bool> runWhile, TimeSpan? timeout = null, Action? onTick = null) =>
             RunPlayerCommands(world, angle, commands, runWhile, timeout, onTick);